fixed an error
since you passed a float to getMoney, used get.uniform instead of get.randint since for whatever reason it would be passed a float sometimes. (maybe because of the forfeit? example: 47$ bet, would be 23.5. i don't think it would like it, too lazy to test it, since i already fixed it) 
the only thing i'm concerned about is the occasional disappearance of players (the game would randomly end despite a player still being available to play) with that said, it's always player 2, when testing with just 3 players. 

nonetheless, running this does NOT get an error anymore (randomly), which personally i'm happy about
i did say to add a stop of 30 rounds, but i managed to get a 41 round simulation
at this point, i don't think it matters, but just wanted to let you know.